BackEnd/requirements.txt
boto3==1.38.44
botocore==1.38.44
Django==5.2.3
djangorestframework==3.16.0
PyJWT==2.10.1
pynamodb==6.1.0
python-dotenv==1.1.1
Requests==2.32.4
nltk==3.8.1